% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proteome_from_fasta.R
\name{read_proteom_from_fasta}
\alias{read_proteom_from_fasta}
\title{Read uniprot FASTA file}
\usage{
read_proteom_from_fasta(filename)
}
\arguments{
\item{filename}{The path to the FASTA file that should be imported.}
}
\value{
A data frame that contains a column with the uniprot ID and a column with the sequence of the protein.
}
\description{
This function is able to read a proteome FASTA file from uniprot.
}
\examples{
\dontrun{
median_normalisation("folder\filename.FASTA")
}
}
