% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fread_tibble.R
\name{clean_fread_tibble}
\alias{clean_fread_tibble}
\title{Fast read, cleaning and conversion}
\usage{
clean_fread_tibble(.filename, ...)
}
\arguments{
\item{.filename}{The path to the file that should be imported.}

\item{...}{Additional arguments for the fread function.}
}
\value{
A data frame (with class tibble) that contains the content of the file.
}
\description{
The functions uses data.tables fast \code{fread} function. The columns of the resulting data.table are made more r-friendly. In the end the data.table is converted to a tibble.
}
\examples{
\dontrun{
clean_fread_tibble("folder\.filename")
}
}
